[simplex2d] fixed overflow

+++ 2-simplex/test2-accum/src/kernels.cuh
@@ -26,9 +26,9 @@
 
 #define OFFSET -0.4999f
 //#define OFFSET 0.5f
-__device__ void work(DTYPE* data, MTYPE* mat, uint2 p, int n, const int a) {
+__device__ void work(DTYPE* data, MTYPE* mat, uint2 p, unsigned int n, const int a) {
     // (1) constant write
-    unsigned long i = (unsigned long)p.y * n + (unsigned long)p.x;
+    unsigned long i = p.y * (unsigned long)n + p.x;
     mat[i] += a;
     // mat[i] = a;
 
@@ -37,9 +37,9 @@ __device__ void work(DTYPE* data, MTYPE* mat, uint2 p, int n, const int a) {
     // mat[p.y * n + p.x] = b;
 }
 
-__device__ void workSubBlock(DTYPE* data, MTYPE* mat, uint2 p, int n, const int a) {
+__device__ void workSubBlock(DTYPE* data, MTYPE* mat, uint2 p, unsigned int n, const int a) {
     // (1) constant write
-    //unsigned long i = (unsigned long)p.y * n + (unsigned long)p.x;
+    // unsigned long i = (unsigned long)p.y * n + (unsigned long)p.x;
     // mat[i] += a;
     // mat[i] = a;
 
@@ -55,7 +55,7 @@ __device__ void workSubBlock(DTYPE* data, MTYPE* mat, uint2 p, int n, const int 
     int subBlockIdx_y = inner_coord_y + blockIdx.y * subBlocksPerGPUblock_x;
 
     const int b = (int)log2f(subBlockIdx_y + 1) + 1;
-    mat[p.y * n + p.x] = b;
+    mat[p.y * (size_t)n + p.x] = b;
 }
 
 // metodo kernel test
@@ -117,50 +117,50 @@ __global__ void kernel_test_DP(const unsigned int n, const unsigned int levelBlo
 }
 
 // O(n^2) number of threads for work (on = original n)
-__global__ void kernelDP_work(int on, int n, MTYPE *data, int offX, int offY){
+__global__ void kernelDP_work(unsigned int on, unsigned int n, MTYPE* data, unsigned int offX, unsigned int offY) {
     // Process data
-    auto p = (uint2) {blockIdx.x*blockDim.x + threadIdx.x, blockIdx.y*blockDim.y + threadIdx.y};
-    //printf("thread at local x=%i  y=%i\n", p.x, p.y);
-    if(p.x >= n || p.y >= n){
-        //printf("discarding thread at local x=%i  y=%i\n", p.x, p.y);
+    auto p = (uint2) { blockIdx.x * blockDim.x + threadIdx.x, blockIdx.y * blockDim.y + threadIdx.y };
+    // printf("thread at local x=%i  y=%i\n", p.x, p.y);
+    if (p.x >= n || p.y >= n) {
+        // printf("discarding thread at local x=%i  y=%i\n", p.x, p.y);
         return;
     }
     p.x = p.x + offX;
     p.y = p.y + offY;
-    //printf("checking thread at global x=%i  y=%i\n", p.x, p.y);
-    if(p.y >= p.x && p.y < on){
-        //printf("work at x=%i  y=%i\n", p.x, p.y);
+    // printf("checking thread at global x=%i  y=%i\n", p.x, p.y);
+    if (p.y >= p.x && p.y < on) {
+        // printf("work at x=%i  y=%i\n", p.x, p.y);
         work(NULL, data, p, on, 1);
     }
 }
 
 // 1 thread does exploration (on = original n)
-__global__ void kernelDP_exp(int on, int n, MTYPE* data, int x0, int y0, int MIN_SIZE){
-    #ifdef DP
-        // 1) stopping case
-        if(n <= MIN_SIZE){
-            dim3 bleaf(BSIZE2D, BSIZE2D), gleaf = dim3((n+bleaf.x-1)/bleaf.x, (n+bleaf.y-1)/bleaf.y, 1);
-            //printf("leaf kernel at x=%i  y=%i   size %i x %i (grid (%i,%i,%i)  block(%i,%i,%i))\n", x0, y0, n, n, gleaf.x, gleaf.y, gleaf.z, bleaf.x, bleaf.y, bleaf.z);
-            kernelDP_work<<<gleaf, bleaf>>>(on, n, data, x0, y0);
-            return;
-        }
-        // 2) explore up and right asynchronously
-        cudaStream_t s1, s2, s3;
-        cudaStreamCreateWithFlags(&s1, cudaStreamNonBlocking);
-        cudaStreamCreateWithFlags(&s2, cudaStreamNonBlocking);
-        cudaStreamCreateWithFlags(&s3, cudaStreamNonBlocking);
-        int subn = (n >> 1) + (n & 1);
-        int n2 = n >> 1;
-        //printf("subn %i\nn2 %i\n", subn, n2);
-        // up
-        kernelDP_exp<<<1,1,0,s1>>>(on, n2, data, x0       , y0       , MIN_SIZE);
-        // bottom right
-        kernelDP_exp<<<1,1,0,s2>>>(on, n2, data, x0 + subn, y0 + subn, MIN_SIZE);
-        // 3) work in the bot middle
-        dim3 bnode(BSIZE2D, BSIZE2D);
-        dim3 gnode = dim3((subn+bnode.x-1)/bnode.x, (subn+bnode.y-1)/bnode.y, 1);
-        //printf("node kernel at x=%i  y=%i   size %i x %i\n", x0, y0+n2, subn, subn);
-        kernelDP_work<<<gnode, bnode, 0, s3>>>(on, subn, data, x0, y0 + n2);
-    #endif
+__global__ void kernelDP_exp(unsigned int on, unsigned int n, MTYPE* data, unsigned int x0, unsigned int y0, unsigned int MIN_SIZE) {
+#ifdef DP
+    // 1) stopping case
+    if (n <= MIN_SIZE) {
+        dim3 bleaf(BSIZE2D, BSIZE2D), gleaf = dim3((n + bleaf.x - 1) / bleaf.x, (n + bleaf.y - 1) / bleaf.y, 1);
+        // printf("leaf kernel at x=%i  y=%i   size %i x %i (grid (%i,%i,%i)  block(%i,%i,%i))\n", x0, y0, n, n, gleaf.x, gleaf.y, gleaf.z, bleaf.x, bleaf.y, bleaf.z);
+        kernelDP_work<<<gleaf, bleaf>>>(on, n, data, x0, y0);
+        return;
+    }
+    // 2) explore up and right asynchronously
+    cudaStream_t s1, s2, s3;
+    cudaStreamCreateWithFlags(&s1, cudaStreamNonBlocking);
+    cudaStreamCreateWithFlags(&s2, cudaStreamNonBlocking);
+    cudaStreamCreateWithFlags(&s3, cudaStreamNonBlocking);
+    int subn = (n >> 1) + (n & 1);
+    int n2 = n >> 1;
+    // printf("subn %i\nn2 %i\n", subn, n2);
+    //  up
+    kernelDP_exp<<<1, 1, 0, s1>>>(on, n2, data, x0, y0, MIN_SIZE);
+    // bottom right
+    kernelDP_exp<<<1, 1, 0, s2>>>(on, n2, data, x0 + subn, y0 + subn, MIN_SIZE);
+    // 3) work in the bot middle
+    dim3 bnode(BSIZE2D, BSIZE2D);
+    dim3 gnode = dim3((subn + bnode.x - 1) / bnode.x, (subn + bnode.y - 1) / bnode.y, 1);
+    // printf("node kernel at x=%i  y=%i   size %i x %i\n", x0, y0+n2, subn, subn);
+    kernelDP_work<<<gnode, bnode, 0, s3>>>(on, subn, data, x0, y0 + n2);
+#endif
 }
 #endif
